+++ packages/animations/test/page_transition_switcher_test.dart
@@ -581,7 +581,7 @@ void main() {
 }
 
 class StatefulTestWidget extends StatefulWidget {
-  const StatefulTestWidget({Key? key}) : super(key: key);
+  const StatefulTestWidget({super.key});
 
   @override
   StatefulTestWidgetState createState() => StatefulTestWidgetState();