+++ packages/video_player/video_player_android/android/src/main/java/io/flutter/plugins/videoplayer/VideoPlayer.java
@@ -56,7 +56,7 @@ final class VideoPlayer {
 
   private final EventChannel eventChannel;
 
-  private final String USER_AGENT = "User-Agent";
+  private static final String USER_AGENT = "User-Agent";
 
   @VisibleForTesting boolean isInitialized = false;